refactor: Code sharing btw. Generic+Aligned via inheritance
