+++ Examples/Detectors/GenericDetector/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_library(
     ActsExamplesDetectorGeneric
     SHARED
-    src/BuildGenericDetector.cpp
     src/GenericDetector.cpp
     src/GenericDetectorElement.cpp
     src/LayerBuilder.cpp
